refactor: converge runtime architecture and engine tooling
- consolidate runtime, editor, hub, plugin, and interface module boundaries
- reorganize plugin audit utilities and Zircon export templates under the tooling surface
- remove obsolete runtime fallback paths and stale compatibility structure
- extend render feature-pass construction with shader authoring contracts and acceptance coverage
- update workspace configuration, CI, validation scripts, fixtures, examples, and regression tests
- shorten deeply nested runtime test paths so Git can stage the workspace reliably on Windows
- split oversized plan status records into focused output documents and refresh architecture documentation
- add the related untracked source, test, documentation, and tooling files required by these changes

+++ zircon_runtime/src/script/vm/runtime/vm_plugin_slot_record.rs
@@ -9,7 +9,7 @@ pub struct VmPluginSlotRecord {
     pub slot: PluginSlotId,
     pub backend_name: String,
     pub state: VmPluginSlotState,
-    pub generation: u64,
+    pub generation: u32,
     pub source: VmPluginPackageSource,
     pub manifest: VmPluginManifest,
     pub management: VmPluginManagementPolicy,
